utils almost sorted
