+++ app/components/events-page/create-event.tsx
@@ -5,7 +5,8 @@ import toast from 'react-hot-toast';
 import { useForm } from 'react-hook-form';
 import { Input } from '../input';
 import { Button } from '../button';
-import { generateDays, generateMonths, generateYears, generateHours, generateMinutes, placeholderImages } from '@/app/lib/utils';
+import { generateDays, generateMonths, generateYears, generateHours, generateMinutes } from '@/app/lib/utils/time';
+import { placeholderImages } from '@/app/lib/utils/events';
 import { useRouter } from 'next/navigation';
 import { ArrowLeftIcon, TrashIcon } from '@heroicons/react/24/outline';
 import { useState, useEffect, useRef } from 'react';
