+++ app/components/events-page/edit-event.tsx
@@ -4,7 +4,8 @@ import Image from 'next/image';
 import toast from 'react-hot-toast';
 import { useState, useEffect, useRef } from 'react';
 import { useForm } from 'react-hook-form';
-import { generateDays, generateMonths, generateYears, generateHours, generateMinutes, placeholderImages } from '@/app/lib/utils';
+import { generateDays, generateMonths, generateYears, generateHours, generateMinutes } from '@/app/lib/utils/time';
+import { placeholderImages } from '@/app/lib/utils/events';
 import { Event, FormData, Registrations, EditEventProps } from '@/app/lib/types';
 import { Input } from '../input';
 import { Button } from '../button';
