+++ app/components/events-page/event-tag-filters.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { EVENT_TAG_TYPES } from "@/app/lib/utils";
+import { EVENT_TAG_TYPES } from "@/app/lib/utils/events";
 
 interface TagButtonsProps {
 	activeTags: number[];
