+++ app/components/input.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "@/app/lib/utils"
+import { cn } from "../lib/utils/general"
 
 export interface InputProps
 	extends React.InputHTMLAttributes<HTMLInputElement> { }
