+++ app/components/table.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "@/app/lib/utils"
+import { cn } from "../lib/utils/general"
 
 const Table = React.forwardRef<
   HTMLTableElement,
