+++ app/events/[id]/event-info.tsx
@@ -4,8 +4,10 @@ import { useParams } from "next/navigation";
 import { Event } from "@/app/lib/types";
 import { useEffect, useState } from "react";
 import Image from "next/image";
-import { base62ToBase16 } from "@/app/lib/uuid-utils";
-import { EVENT_TAG_TYPES, returnLogo, formatDateString } from "@/app/lib/utils";
+import { base62ToBase16 } from "@/app/lib/utils/type-manipulation";
+import { returnLogo } from "@/app/lib/utils";
+import { EVENT_TAG_TYPES } from "@/app/lib/utils/events";
+import { formatDateString } from "@/app/lib/utils/time";
 import { LockClosedIcon, ArrowRightIcon } from "@heroicons/react/24/outline";
 import { useSession } from "next-auth/react";
 import toast from "react-hot-toast";
