+++ app/lib/data.ts
@@ -1,6 +1,10 @@
 import { sql } from '@vercel/postgres';
 import { SQLEvent, ContactFormInput, SocietyRegisterFormData, UserRegisterFormData, SQLRegistrations, OrganiserAccountEditFormData, CompanyRegisterFormData, InsertTokenResult, EventRegistrationEmail } from './types';
-import { convertSQLEventToEvent, formatDOB, selectUniversity, capitalize, convertSQLRegistrationsToRegistrations, capitalizeFirst, FallbackStatistics } from './utils';
+import { FallbackStatistics } from './utils';
+import { selectUniversity } from './utils/events';
+import { formatDOB } from './utils/events';
+import { capitalize, capitalizeFirst } from './utils/general';
+import { convertSQLEventToEvent, convertSQLRegistrationsToRegistrations } from './utils/type-manipulation';
 import bcrypt from 'bcrypt';
 import { Tag } from './types';
 import { getRedisClient } from './config';
